chore: Update dependencies and add DaisyUI
This commit updates the dependencies in the `package.json` file and adds the `daisyui` package. DaisyUI is a utility-first CSS framework for rapidly building custom designs. It provides a set of pre-built components and utilities that can be easily customized. This update will enhance the styling capabilities of the project and enable the use of DaisyUI features.