chore: Add svelte-hamburgers package and update dependencies
The svelte-hamburgers package has been added as a new dependency to the project. Additionally, the package-lock.json and package.json files have been updated to include the latest version of svelte-hamburgers (4.2.0) and other dependencies. This change ensures that the project has access to the svelte-hamburgers library and is using up-to-date dependencies.